requirements.txt
# Core dependencies
python-dotenv==1.0.0
gradio==4.44.0

# LangChain ecosystem
langchain==0.2.16
langchain-openai==0.1.23
langchain-chroma==0.1.4
langchain-community==0.2.16

# Document processing
pypdf==4.3.1
python-docx==1.1.2
docx2txt==0.8
pandas==2.2.2
openpyxl==3.1.5

# Vector store and embeddings
chromadb==0.5.5

# Visualization (optional, for debugging)
matplotlib==3.9.2
plotly==5.24.1
scikit-learn==1.5.1

# Additional utilities
numpy==2.0.1
pathlib2==2.3.7.post1

# Optional: For enhanced text processing
tiktoken==0.7.0